requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
crewai
langchain
langchain-openai
beautifulsoup4
requests
python-multipart
aiohttp
pandas
numpy
scikit-learn
python-jose
passlib
bcrypt
sqlalchemy
alembic
psycopg2-binary
pytest
httpx
pydantic-email
fake-useragent
google-generativeai
email-validator 